+++ docx4j-core-tests/src/test/java/org/docx4j/model/PropertyResolverTestUtils.java
@@ -49,7 +49,7 @@ public class PropertyResolverTestUtils {
 		
 	}
 	
-	private static WordprocessingMLPackage createdPkgWithStyles(Styles styles) throws InvalidFormatException {
+	static WordprocessingMLPackage createdPkgWithStyles(Styles styles) throws InvalidFormatException {
 		
 		// Create a package
 		WordprocessingMLPackage wmlPack = new WordprocessingMLPackage();